chore: 분리된 타입 파일 삭제
