--- src/page/qna-detail/components/Answers/AnswerList.types.ts
@@ -1,9 +1,0 @@
-export interface AnswerProps {
-  createdby: string
-  questionId: number
-  isMyAnswer: boolean
-}
-
-export interface NoAnswerProps {
-  isMyAnswer: boolean
-}
