+++ src/page/qna-detail/components/Answers/DayBox.tsx
@@ -2,7 +2,11 @@
 
 import { getKorRelativeTime } from "@/util/getDate"
 import dayjs from "dayjs"
-import type { DayBoxProps } from "./DayBox.types"
+import type { Answer } from "@/interfaces/answer"
+
+export type DayBoxProps = {
+  answer: Answer
+}
 
 const DayBox: React.FC<DayBoxProps> = ({ answer }) => {
   const now = dayjs().format()
