--- src/page/qna-detail/components/Answers/DayBox.types.ts
@@ -1,5 +1,0 @@
-import type { Answer } from "@/interfaces/answer"
-
-export type DayBoxProps = {
-  answer: Answer
-}
