+++ src/page/qna-detail/components/Answers/HandleAnswerBox.tsx
@@ -3,7 +3,12 @@
 import useHandleMyAnswer from "../../hooks/useHandleMyAnswer"
 import SuccessModalContent from "../SuccessModalContent"
 import { buttonMessage, successMessage } from "@/constants/message"
-import type { HandleAnswerProps } from "./HandleAnswerBox.types"
+import type { Answer } from "@/interfaces/answer"
+
+export type HandleAnswerProps = {
+  answer: Answer
+  createdby?: string
+}
 
 const HandleAnswerBox: React.FC<HandleAnswerProps> = ({
   answer,
