--- src/page/qna-detail/components/Answers/HandleAnswerBox.types.ts
@@ -1,6 +1,0 @@
-import type { Answer } from "@/interfaces/answer"
-
-export type HandleAnswerProps = {
-  answer: Answer
-  createdby?: string
-}
