--- src/page/qna-detail/components/Answers/OneAnswer.types.ts
@@ -1,6 +1,0 @@
-import { Answer } from "@/interfaces/answer"
-
-export interface OneAnswerProps {
-  answer: Answer
-  createdby: string
-}
