--- src/page/qna-detail/components/Answers/VoteBox.types.ts
@@ -1,6 +1,0 @@
-import type { Answer } from "@/interfaces/answer"
-
-export type VoteBoxProps = {
-  userId?: number
-  answer: Answer
-}
