--- src/page/qna-detail/components/Markdown/MdEditor.types.ts
@@ -1,7 +1,0 @@
-import type { Editor } from "@toast-ui/react-editor"
-import type { RefObject } from "react"
-
-export interface EditorProps {
-  previous?: string
-  editorRef: RefObject<Editor>
-}
