--- src/page/qna-detail/components/Markdown/MdViewer.types.ts
@@ -1,3 +1,0 @@
-export interface ViewerProps {
-  content: string
-}
