--- src/page/qna-detail/components/MyAnswer.types.ts
@@ -1,3 +1,0 @@
-export interface MyAnswerProps {
-  questionId: number
-}
