--- src/page/qna-detail/components/Question/HandleQuestionBox.types.ts
@@ -1,5 +1,0 @@
-import type { Question } from "@/interfaces/question"
-
-export interface HandleQuestionBoxProps {
-  question: Question
-}
