+++ src/page/qna-detail/components/Question/WriterBox.tsx
@@ -3,7 +3,11 @@
 import { Icons } from "@/components/icons/Icons"
 import Image from "next/image"
 import { useRouter } from "next/navigation"
-import type { WriterBoxProps } from "./WriterBox.types"
+import type { Question } from "@/interfaces/question"
+
+export interface WriterBoxProps {
+  question: Question
+}
 
 const WriterBox: React.FC<WriterBoxProps> = ({ question }) => {
   const router = useRouter()
