--- src/page/qna-detail/components/Question/WriterBox.types.ts
@@ -1,5 +1,0 @@
-import type { Question } from "@/interfaces/question"
-
-export interface WriterBoxProps {
-  question: Question
-}
