--- src/page/qna-detail/components/SuccessModalContent.types.ts
@@ -1,10 +1,0 @@
-import { CSSProperties } from "react"
-
-export interface SuccesssModalProps {
-  message: string
-}
-
-export interface IconProps {
-  message: string
-  style?: CSSProperties | undefined
-}
