--- src/page/qna-detail/hooks/useQnADetail.types.ts
@@ -1,4 +1,0 @@
-export interface SubmitValueProps {
-  questionId: number
-  submitValue: string | undefined
-}
